Add padding to infinite scroll view indicator
