+++ src/components/Explore/Explore.js
@@ -16,8 +16,6 @@ const mapHeight = height - 450;
 
 const Explore = ( ): Node => {
   const {
-    exploreList,
-    loadingExplore,
     exploreFilters
   } = useContext( ExploreContext );
   const taxonId = exploreFilters ? exploreFilters.taxon_id : null;
@@ -26,12 +24,6 @@ const Explore = ( ): Node => {
     <ViewWithFooter>
       {taxonId !== null && (
         <ObservationViews
-          loading={loadingExplore}
-          localObservations={{
-            observationList: exploreList,
-            unuploadedObsList: [],
-            allObsToUpload: []
-          }}
           taxonId={taxonId}
           testID="Explore.observations"
           mapHeight={mapHeight}
