+++ src/components/Observations/InfiniteScrollFooter.js
@@ -6,7 +6,7 @@ import React from "react";
 import { ActivityIndicator } from "react-native";
 
 const InfiniteScrollFooter = ( ): Node => (
-  <View className="h-32 border border-border pt-10">
+  <View className="h-32 border border-border py-16">
     <ActivityIndicator />
   </View>
 );
